Add Swagger UI to api/index.html and check in required assets
This adds the required files/libraries to run swagger UI on
api/index.html. I didn't immediately see any alternatives or easy way
to load it, and this will at least keep everything the same while we
build out the site more.

The only question left is if this will work on github-pages in an
actual deployment. It _does_ work locally with jekyll and the github
ages gem.