Fix overflowing parameters in swagger-ui documentation
This adds a custom swagger.scss asset file to the repo that will be
used on the /api/index.html -- swagger UI -- page to prevent text from
overflowing in the parameters section of the endpoints.

I can't exactly tell the root cause of *why* it started overflowing,
but this will be a preventative fix so that we can make pretty much
any updates to the descriptions without needing to worry about
entering newlines, etc.